Create FUNDING.yml
I'm writing to ask you to support me and my coding project or buy me a cup of coffee. Just a small donation of $10 can help me to allows programmers to build programs, such as websites and apps. Your donation will go toward impacting others to be generated in terms of coding. [When possible, add a personal connection to tie the donor to the cause.
paypal number is:+254759873854 /@denzelamani /denzelamani20@gmail.com

Signed-off-by: @denzelamani <125563552+DenzelAmani@users.noreply.github.com>